Fixups to enable deployment to appspot.
Clojure can't ask for the system classloader (due to security restrictions).
Compojure can't store a clojure.lang.Ref in HttpSession (because it's not Serializable).

Plus a clean target I realized I needed after updating the Compojure jar and forgetting I'd also need to rebuild my app classes (as the change I made was in a macro).